Breaking everything up. Also added a cool table that changes with the line movement
